+++ tests/test_ids.py
@@ -1,8 +1,20 @@
 import pytest
 
-from ome_types import from_xml
+from ome_types import from_xml, model
 from ome_types._mixins import _ids
 from ome_types.model import Line, Rectangle
+
+
+def test_no_id() -> None:
+    """Test that ids are optional, and auto-increment."""
+    i = model.Instrument(id=20)  # type: ignore
+    assert i.id == "Instrument:20"
+    i2 = model.Instrument()  # type: ignore
+    assert i2.id == "Instrument:21"
+
+    # but validation still works
+    with pytest.warns(match="Casting invalid InstrumentID"):
+        model.Instrument(id="nonsense")
 
 
 def test_shape_ids(monkeypatch: "pytest.MonkeyPatch") -> None:
